requirements.txt
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf
python-dotenv
streamlit